requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0 
torchvision==0.7.0
pytorch-ignite==0.4.9
imgaug==0.4.0
